+++ open_instruct/utils.py
@@ -11,19 +11,39 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+# isort: off
+import os
 
+os.environ["NCCL_CUMEM_ENABLE"] = "0"  # NOQA
+try:
+    import deepspeed
+    from deepspeed.runtime.zero.partition_parameters import ZeroParamStatus
+    from deepspeed.runtime.zero.offload_config import OffloadDeviceEnum
+
+    # @vwxyzjn: when importing on CPU-only machines, we get the following error:
+    # RuntimeError: 0 active drivers ([]). There should only be one.
+    # so we need to catch the exception and do nothing
+    # https://github.com/deepspeedai/DeepSpeed/issues/7028
+except Exception:
+    pass
+# isort: on
 import dataclasses
 import functools
 import json
 import logging
 import os
+import random
 import shutil
+import socket
 import subprocess
 import sys
 import time
 from dataclasses import dataclass
 from typing import Any, List, NewType, Optional, Tuple, Union
+from ctypes import CDLL, POINTER, Structure, c_char_p, c_int, c_ulong, c_void_p
 
+import numpy as np
+import ray
 import requests
 from accelerate.logging import get_logger
 from datasets import DatasetDict, concatenate_datasets, load_dataset, load_from_disk
@@ -31,6 +51,7 @@ from datasets.builder import DatasetGenerationError
 from dateutil import parser
 from huggingface_hub import HfApi
 from rich.pretty import pprint
+import torch
 from transformers import MODEL_FOR_CAUSAL_LM_MAPPING, HfArgumentParser
 
 MODEL_CONFIG_CLASSES = list(MODEL_FOR_CAUSAL_LM_MAPPING.keys())
@@ -1036,3 +1057,238 @@ def upload_metadata_to_hf(
         repo_type="dataset",
     )
     os.remove("tmp.json")
+
+
+# ----------------------------------------------------------------------------
+# Ray utilities
+
+
+
+def get_train_ds_config(
+    offload,
+    adam_offload=False,
+    stage=0,
+    bf16=True,
+    max_norm=1.0,
+    zpg=8,
+    grad_accum_dtype=None,
+    disable_trace_cache=True,
+):
+    device = "cpu" if offload else "none"
+    zero_opt_dict = {
+        "stage": stage,
+        "offload_param": {"device": device},
+        "offload_optimizer": {
+            "device": "cpu" if adam_offload else "none",
+            "pin_memory": True,
+        },
+        "sub_group_size": "auto",
+        "stage3_max_live_parameters": "auto",
+        "stage3_max_reuse_distance": "auto",
+        "stage3_param_persistence_threshold": "auto",
+        "stage3_prefetch_bucket_size": "auto",
+        "reduce_bucket_size": "auto",
+        # # ZeRO++
+        # "zero_hpz_partition_size": zpg,
+        # "zero_quantized_weights": False,
+        # "zero_quantized_gradients": False,
+    }
+    if disable_trace_cache:
+        zero_opt_dict["stage3_prefetch_bucket_size"] = 0
+        zero_opt_dict["stage3_max_live_parameters"] = 0
+        zero_opt_dict["stage3_max_reuse_distance"] = 0
+
+    return {
+        "steps_per_print": 100,
+        "zero_optimization": zero_opt_dict,
+        "bf16": {
+            "enabled": bf16,
+        },
+        "gradient_clipping": max_norm,
+        "prescale_gradients": False,
+        "wall_clock_breakdown": False,
+        "data_types": {"grad_accum_dtype": grad_accum_dtype if grad_accum_dtype else "fp32"},
+    }
+
+
+def get_eval_ds_config(
+    offload,
+    stage=0,
+    bf16=True,
+):
+    zero_opt_dict = {
+        "stage": stage,
+        "stage3_param_persistence_threshold": "auto",
+        "offload_param": {
+            "device": "cpu" if offload else "none",
+            "pin_memory": True,
+        },
+    }
+    return {
+        "steps_per_print": 100,
+        "zero_optimization": zero_opt_dict,
+        "bf16": {
+            "enabled": bf16,
+        },
+        "prescale_gradients": False,
+        "wall_clock_breakdown": False,
+    }
+
+
+def get_optimizer_grouped_parameters(
+    model: torch.nn.Module,
+    weight_decay: float,
+    no_decay_name_list=["bias", "layer_norm.weight", "layernorm.weight", "norm.weight", "ln_f.weight"],
+):
+    optimizer_grouped_parameters = [
+        {
+            "params": [
+                p
+                for n, p in model.named_parameters()
+                if (not any(nd in n for nd in no_decay_name_list) and p.requires_grad)
+            ],
+            "weight_decay": weight_decay,
+        },
+        {
+            "params": [
+                p
+                for n, p in model.named_parameters()
+                if (any(nd in n for nd in no_decay_name_list) and p.requires_grad)
+            ],
+            "weight_decay": 0.0,
+        },
+    ]
+    return optimizer_grouped_parameters
+
+
+def _z3_params_to_fetch(param_list):
+    return [p for p in param_list if hasattr(p, "ds_id") and p.ds_status == ZeroParamStatus.NOT_AVAILABLE]
+
+
+
+
+def get_ray_address() -> Optional[str]:
+    """Get the Ray address from the environment variable."""
+    return os.environ.get("RAY_ADDRESS")
+
+
+
+class RayProcess:
+    def __init__(self, world_size, rank, local_rank, master_addr, master_port):
+        logging.basicConfig(
+            format="%(asctime)s %(levelname)-8s %(message)s",
+            level=logging.INFO,
+            datefmt="%Y-%m-%d %H:%M:%S",
+        )
+        self.world_size = world_size
+        self.rank = rank
+        self.local_rank = local_rank
+        self.master_addr = master_addr if master_addr else self.get_current_node_ip()
+        self.master_port = master_port if master_port else self.get_free_port()
+        os.environ["MASTER_ADDR"] = self.master_addr
+        os.environ["MASTER_PORT"] = str(self.master_port)
+        os.environ["WORLD_SIZE"] = str(self.world_size)
+        os.environ["RANK"] = str(self.rank)
+        # NOTE: Ray will automatically set the CUDA_VISIBLE_DEVICES
+        # environment variable for each actor, so always set device to 0
+        # os.environ["LOCAL_RANK"] = str(self._local_rank)
+        os.environ["LOCAL_RANK"] = "0"
+        random.seed(self.rank)
+        np.random.seed(self.rank)
+        torch.manual_seed(self.rank)
+
+    @staticmethod
+    def get_current_node_ip():
+        address = ray._private.services.get_node_ip_address()
+        # strip ipv6 address
+        return address.strip("[]")
+
+    @staticmethod
+    def get_free_port():
+        with socket.socket() as sock:
+            sock.bind(("", 0))
+            return sock.getsockname()[1]
+
+    def get_master_addr_port(self):
+        return self.master_addr, self.master_port
+
+    def empty_cache(self) -> None:
+        torch.cuda.empty_cache()
+
+
+    def _set_numa_affinity(self, rank):
+        def local_rank_to_real_gpu_id(local_rank):
+            cuda_visible_devices = [
+                int(x) for x in os.environ.get("CUDA_VISIBLE_DEVICES", "0,1,2,3,4,5,6,7").split(",")
+            ]
+            return cuda_visible_devices[local_rank]
+
+        rank = local_rank_to_real_gpu_id(rank)
+
+        global _SET_AFFINITY
+        if _SET_AFFINITY:
+            return
+
+        from ctypes.util import find_library
+
+        class bitmask_t(Structure):
+            _fields_ = [
+                ("size", c_ulong),
+                ("maskp", POINTER(c_ulong)),
+            ]
+
+        LIBNUMA = CDLL(find_library("numa"))
+        LIBNUMA.numa_parse_nodestring.argtypes = [c_char_p]
+        LIBNUMA.numa_parse_nodestring.restype = POINTER(bitmask_t)
+        LIBNUMA.numa_run_on_node_mask.argtypes = [POINTER(bitmask_t)]
+        LIBNUMA.numa_run_on_node_mask.restype = c_int
+        LIBNUMA.numa_set_membind.argtypes = [POINTER(bitmask_t)]
+        LIBNUMA.numa_set_membind.restype = c_void_p
+        LIBNUMA.numa_num_configured_nodes.argtypes = []
+        LIBNUMA.numa_num_configured_nodes.restype = c_int
+
+        def numa_bind(nid: int):
+            bitmask = LIBNUMA.numa_parse_nodestring(bytes(str(nid), "ascii"))
+            LIBNUMA.numa_run_on_node_mask(bitmask)
+            LIBNUMA.numa_set_membind(bitmask)
+
+        numa_nodes = LIBNUMA.numa_num_configured_nodes()
+        num_gpu_pre_numa_node = 8 // numa_nodes
+        numa_bind(self._local_rank // num_gpu_pre_numa_node)
+        _SET_AFFINITY = True
+
+    def offload_to_cpu(self, model, pin_memory=True, non_blocking=True):
+        """This function guaratees the memory are all released (only torch context cache <100M will remain)."""
+        self._set_numa_affinity(torch.distributed.get_rank() % torch.cuda.device_count())
+
+        if model.zero_optimization_stage() == 3:
+            from deepspeed.runtime.zero.offload_config import OffloadStateTypeEnum
+
+            model.optimizer.offload_states(
+                include=[
+                    OffloadStateTypeEnum.optim_states,
+                    OffloadStateTypeEnum.contiguous_grad_buffer,
+                    OffloadStateTypeEnum.hp_params,
+                    # OffloadStateTypeEnum.lp_grads,
+                    # OffloadStateTypeEnum.lp_params, # dangerous
+                ],
+                device=OffloadDeviceEnum.cpu,
+                pin_memory=pin_memory,
+                non_blocking=non_blocking,
+            )
+            torch.cuda.synchronize()
+            return
+
+        raise NotImplementedError("Zero stage 2 is not supported yet")
+
+    def backload_to_gpu(self, model, non_blocking=True):
+        # NOTE: this function reloads the weights, ensuring the calculation
+        if model.zero_optimization_stage() == 3:
+            model.reload_states(non_blocking=non_blocking)
+            torch.cuda.synchronize()
+            return
+
+        raise NotImplementedError("Zero stage 2 is not supported yet")
+
+
+
